requirements.txt
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dnspython==2.6.1
idna==3.6
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pluggy==1.4.0
pymongo==4.6.3
pytest==8.1.1
python-dotenv==1.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
urllib3==2.2.1
